% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcademy-package.R
\docType{data}
\name{abdc}
\alias{abdc}
\title{Australian Business Deans Council Rankings}
\format{tibble}
\source{
Australian Business Deans Council
}
\usage{
abdc
}
\description{
Journal rankings according to the Australian Business Deans Council (ABDC)
}
\examples{
abdc

}
\keyword{datasets}
