% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{citations}
\alias{citations}
\title{Obtain citations based on DOIs using CrossRef data}
\usage{
citations(doi)
}
\arguments{
\item{doi}{Unquoted column containing DOIs}
}
\value{
A vector of citation counts from CrossRef OpenURL
}
\description{
Return CrossRef citation information for items in
data frame of publications using DOIs.
}
\examples{

\dontrun{
njtpubs \%>\%
  mutate(cr_cites = citations(doi))
}

}
\author{
Rob J Hyndman
}
