% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{compute_hindex}
\alias{compute_hindex}
\title{Compute h-index given papers and citations}
\usage{
compute_hindex(citations)
}
\arguments{
\item{citations}{List of citations for all publications}
}
\value{
numerical value
}
\description{
Take a numerical vector of citation counts and return the h-index
}
\examples{
\dontrun{
compute_hindex(read_scholar("XSyW00YAAAAJ")$cites)
}

}
