% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-core.R
\docType{data}
\name{core}
\alias{core}
\alias{core_journals}
\title{CORE (Computing Research and Education) lists of conference and journal rankings}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 972 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 639 rows and 4 columns.
}
\source{
\url{https://www.core.edu.au/conference-portal}
\url{}
}
\usage{
core

core_journals
}
\description{
Two datasets are provided: \code{core} and \code{core_journals}, which contains lists of
conference and journal rankings respectively, according to the CORE executive committee.
These are used in \code{\link[=rank_core]{rank_core()}}. The details of the CORE organisation,
and its procedure for ranking is provided below.
}
\details{
CORE is an association of university departments of computer science in
Australia and New Zealand. Prior to 2004 it was known as the Computer
Science Association, CSA.

The CORE Conference Ranking provides assessments of major conferences in the
computing disciplines. The rankings are managed by the CORE Executive
Committee, with periodic rounds for submission of requests for addition or
reranking of conferences. Decisions are made by academic committees based
on objective data requested as part of the submission process. Conference
rankings are determined by a mix of indicators, including citation rates,
paper submission and acceptance rates, and the visibility and research
track record of the key people hosting the conference and managing its
technical program. A more detailed statement categorizing the ranks A*, A,
B, and C can be found \href{http://bit.ly/core-rankings}{here}.

Format of \code{core}: A data frame with 972 observations and two variables:
\itemize{
\item{\verb{title:}}{Title of the conference}
\item{\verb{rank:}}{Conferences are assigned to one of the following categories:
\itemize{
\item{A*: flagship conference, a leading venue in a discipline area}
\item{A: excellent conference, and highly respected in a discipline area}
\item{B: good conference, and well regarded in a discipline area}
\item{C: other ranked conference venues that meet minimum standards}
}
}
}
Format of \code{core_journals}: A data frame with 639 observations and five variables:
\itemize{
\item{\verb{title:}}{Title of the journal}
\item{\code{field_of_research}: }{Field of Research Code as provided by the
Australian Bureau of Statistics}
\item{\code{issn}: }{International Standard Serial Number}
\item{\code{rank}: }{In order of best to lowest rank: A*, A, B, or C}
}
}
\examples{
core
core_journals
}
\keyword{datasets}
