% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-era2010.R
\docType{data}
\name{era2010}
\alias{era2010}
\title{ERA2010 Journal List}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20712 rows and 5 columns.
}
\source{
\url{https://www.righttoknow.org.au/request/journal_list_relating_to_the_201}
}
\usage{
data(era2010)
}
\description{
This is a dataset that contains the list of journal rankings from
the ARC Excellence in Research for Australia 2010 round.
}
\details{
Format: a data frame with 20712 rows and the following 4 variables:
\itemize{
\item{\code{eraid}: }{ERA ID of the journal}
\item{\code{journal}: }{Title of the journal}
\item{\code{issn}: }{International Standard Serial Number}
\item{\code{field_of_research}: }{Field of Research Code as provided by the
Australian Bureau of Statistics}
\item{\code{rank}: }{In order of best to lowest rank: A*, A, B, or C}
}
}
\examples{
library(dplyr)
era2010 \%>\%
  filter(field_of_research == "0104") \%>\%
  arrange(rank)
}
\keyword{datasets}
