% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_altmetrics}
\alias{get_altmetrics}
\title{Find Altmetrics}
\usage{
get_altmetrics(data, doi)
}
\arguments{
\item{data}{A data frame or tibble containing a bibliography.}

\item{doi}{The column containing DOI values}
}
\value{
A tibble of altmetrics
}
\description{
Get a tibble of all altmetrics given a list of DOIs
}
\examples{
read_orcid("0000-0002-2140-5352") \%>\%
  altmetrics(doi)

}
