% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{match_citations}
\alias{match_citations}
\title{Match list of citations with Google Scholar citation data}
\usage{
match_citations(mypubs, id)
}
\arguments{
\item{mypubs}{A data frame of publications.}

\item{id}{Google Scholar ID as a text string.}
}
\value{
A data frame of matched and non-matched publications from both inputs.
}
\description{
Add Google Scholar citation information for items in file of publications.
Uses fuzzy matching on title. Keeps non-matches for manual curation.
}
\examples{

\dontrun{

mypubs <- read_pubmed("Rob Hyndman")
matchedPubs <- match_citations(myPubs, "vamErfkAAAAJ")
}

}
\author{
Belinda K Fabian
}
