% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-monash.R
\docType{data}
\name{monash}
\alias{monash}
\title{Monash Business School Journal Quality List}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4614 rows and 2 columns.
}
\source{
Monash Business School
}
\usage{
data(monash)
}
\description{
This is a dataset that contains the list of quality journal rankings from the
Monash Business School. In most cases, it follows
ABDC with A* equal to Group 1 and A equal to Group 2.
The "Group 1+" category contains a small set of the highest rank journals.
}
\details{
Format: a data frame with \code{NROW(monash)} observations on the following 2 variables:
\describe{
\item{\code{title}: }{Title of the journal}
\item{\code{rank}: }{In order of best to lowest rank: Group 1+, Group 1, Group 2}
}
}
\examples{
library(dplyr)
library(stringr)
monash |>
  filter(str_detect(title, "Statist")) |>
  arrange(rank)
}
\keyword{datasets}
