% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-njtpubs.R
\docType{data}
\name{njtpubs}
\alias{njtpubs}
\title{Publications of Nick Tierney}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 8 columns.
}
\source{
\url{https://orcid.org/0000-0003-1460-8722}
}
\usage{
data(njtpubs)
}
\description{
This is a dataset that contains the publications of Nick Tierney available
on ORCID at 20 February 2020.
}
\details{
A data frame with the following variables
\describe{
\item{\code{journal}: }{Title of the journal}
\item{\code{title}: }{Title of the paper}
\item{\code{year}: }{Year of publication}
\item{\code{volume}: }{Volume of the journal}
\item{\code{issue}: }{Issue of the journal}
\item{\code{pages}: }{Page numbers}
\item{\code{type}: }{Type of publications}
\item{\code{doi}: }{Digital Object Identifier}
}
}
\examples{
njtpubs
}
\keyword{datasets}
