% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pdf_emails.R
\name{parse_pdf_emails}
\alias{parse_pdf_emails}
\title{Parsing Emails}
\usage{
parse_pdf_emails(folder)
}
\arguments{
\item{folder}{The path to the folder where all emails are stored.}
}
\value{
A tibble.
}
\description{
This function parses your emails stored as PDF files in a folder.
It returns a data frame with from, to, date, subject and content fields.
Note: some emails if "saved as PDFs" may not be parsed correctly.
}
\examples{
\dontrun{
emails <- parse_pdf_emails("Some/Folder/Path/Here")
}

}
