% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePDFEmails.R
\name{parsePDFEmails}
\alias{parsePDFEmails}
\title{Parsing PDF Emails}
\usage{
parsePDFEmails(folder)
}
\arguments{
\item{folder}{The path to the folder where all emails are.}
}
\value{
The parsed dataframe in-memory
}
\description{
This function parses your emails stored as PDFs files in a folder
and produces an RDS file in the same directory. It contains a dataframe with
from, to, daet, subject and content fields.
Note: some emails if "saved as PDFs" may not be parsed correctly.
}
\examples{
\dontrun{
emails <- parsePDFEmails("Some/Folder/Path/Here")
}

}
