% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{rank_journal}
\alias{rank_journal}
\alias{rank_abdc}
\alias{rank_scimagojr}
\alias{rank_core}
\title{Find rankings of journals from the ABDC, core, or scimagojr lists.}
\usage{
rank_abdc(journal)

rank_scimagojr(journal)

rank_core(journal)
}
\arguments{
\item{journal}{A character vector containing journal names.}
}
\value{
A character vector of the same length as \code{journal} containing
rankings.
}
\description{
Fuzzy matching is used to find the requested journal from the ABDC, CORE,
or scimagojr lists. For more information on each of these, see:
\itemize{
\item{\code{?abdc}: }{for more information on the abdc list}
\item{\code{?core}: }{for more information on the core list}
\item{\code{?scimagojr}: }{for more information on the scimagojr list}
}
}
\examples{
rank_abdc("Annals of statistics")
rank_core("Annals of statistics")
rank_scimagojr("Annals of statistics")
}
\author{
Rob J Hyndman, Nicholas Tierney
}
