% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{rank_scimago}
\alias{rank_scimago}
\alias{rank_abdc}
\alias{rank_era2010}
\alias{rank_core}
\alias{rank_monash}
\alias{journal_ranking}
\title{Find rankings of journals from the ABDC, ERA2010, CORE, SCImago or Monash lists.}
\usage{
rank_abdc(title, fuzzy = TRUE, warning = FALSE)

rank_era2010(title, fuzzy = TRUE, warning = FALSE)

rank_scimago(title, fuzzy = TRUE, warning = FALSE)

rank_core(title, fuzzy = TRUE, warning = FALSE)

rank_monash(title, fuzzy = TRUE, warning = FALSE)

journal_ranking(
  title,
  source = c("all", "abdc", "era2010", "core", "scimago", "monash"),
  fuzzy = TRUE,
  only_best = FALSE,
  return_dist = FALSE,
  ...
)
}
\arguments{
\item{title}{A character vector containing (partial) journal names.}

\item{fuzzy}{Should fuzzy matching be used. If \code{FALSE}, partial exact matching is used.
Otherwise, full fuzzy matching is used.}

\item{warning}{A logical value indicating whether to return warnings when \code{title} contains
missing values.}

\item{source}{A character string indicating which ranking data base to use. Default \code{"all"}.}

\item{only_best}{A logical variable. If \code{TRUE}, only returns the best match found.}

\item{return_dist}{A logical variable. If \code{TRUE}, returns the distance between the \code{title} and the matches found.}

\item{...}{Other arguments are passed to \code{agrepl} (if \code{fuzzy} is \code{TRUE}), or \code{grepl} otherwise.}
}
\value{
The \code{rank_xxx()} functions return a character vector of the same length as \code{title} containing
the rankings from the specified \code{source} database. The \code{journal_ranking()} function returns a tibble
containing the matching journal titles and associated rankings from the specified database.
}
\description{
Fuzzy matching is used to find the requested journal from the ABDC, ERA2010, CORE,
SCImago or Monash lists. For more information on each of these, see:
\itemize{
\item{\cite{\link{abdc}}: }{for more information on the ABDC list}
\item{\cite{\link{era2010}}: }{for more information on the ERA2010 list}
\item{\cite{\link{core}}: }{for more information on the CORE list}
\item{\cite{\link{scimago}}: }{for more information on the SCImago list}
\item{\cite{\link{monash}}: }{for more information on the Monash list}
}
}
\examples{
# Return ranking for individual journals or conferences
rank_abdc("Annals of Statistics")
rank_era2010("Biometrika")
rank_core("International Conference on Machine Learning")
rank_scimago("International Journal of Forecasting")
rank_monash("Annals")

# Add rankings to a data frame of publications
library(dplyr)
njtpubs \%>\%
  mutate(scimago = rank_scimago(journal, warning = FALSE))

# Return rankings from all sources for journals that match a search string
journal_ranking("Forecasting")
}
\author{
Rob J Hyndman
}
