% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{ranking}
\alias{ranking}
\title{Find rankings of journals from Scimago, Core or ABDC. Fuzzy matching is used to
find the requested journal.}
\usage{
ranking(journal, source)
}
\arguments{
\item{journal}{A character vector containing journal names.}

\item{source}{Which journal rankings list to use?}
}
\value{
A character vector of the same length as \code{journal} containing
rankings.
}
\description{
Find rankings of journals from Scimago, Core or ABDC. Fuzzy matching is used to
find the requested journal.
}
\examples{

\dontrun{
mypubs <- read_pubmed("Rob Hyndman")
mypubs <- mypubs \%>\%
  mutate(
    abdc_ranking = ranking(journal, source="abdc"),
    core_ranking = ranking(journal, source="core"),
    scimago_ranking = ranking(journal, source="scimagojr")
  )
}
}
\author{
Rob J Hyndman, Nicholas Tierney
}
