% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcademy-package.R
\docType{package}
\name{rcademy-package}
\alias{rcademy}
\alias{rcademy-package}
\title{rcademy: Tools to assist with academic promotions}
\description{
Ideas and tools to help with preparing documentation for promotions at universities.
}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})

}
\keyword{internal}
