% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_altmetrics}
\alias{read_altmetrics}
\title{Read Altmetrics}
\usage{
read_altmetrics(doi_list)
}
\arguments{
\item{doi_list}{A list of DOI strings for which to return a tibble of Altmetrics}
}
\description{
Get a tibble of all altmetrics given a list of DOIs
}
\examples{
read_altmetrics(list(c("10.1038/nature09210","10.1126/science.1187820")))

}
