% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storeStudents.R
\name{storeStudents}
\alias{storeStudents}
\title{Students parser function}
\usage{
storeStudents(
  firstName,
  familyName,
  degreeType,
  degreeName,
  institution,
  startDate,
  endDate,
  completed = FALSE,
  thesisTitle,
  mainSupervisor,
  adjuctSupervisors,
  csvFilePath,
  csvFileName
)
}
\arguments{
\item{firstName}{The first name of the student.}

\item{familyName}{The student's surname or family name.}

\item{degreeType}{The type of degree: PhD, Master, Honours, and so on.}

\item{degreeName}{The name of the degree.}

\item{institution}{The institution where the student is working.}

\item{startDate}{The start date of the degree.}

\item{endDate}{End date. If it is not completed, a "Expected" will be added before it.}

\item{completed}{False if the student hasn't completed the degree.}

\item{thesisTitle}{The name of the thesis.}

\item{mainSupervisor}{The main supervisor. If there are several, they should be split by comma.}

\item{adjuctSupervisors}{Other supervisors, split by comma.}

\item{csvFilePath}{The path to where the file is stored.}

\item{csvFileName}{The name of the file, with csv extension.}
}
\value{
The in-memory object of the CSV file, with all the students and the updates/adds performed.
}
\description{
This function will allow you to add your students (HDRs, Honours), as well as their basic information. The result will be stored in a CSV file. If it doesn't exist, this creates the file. When you create a new student you need all arguments, but you can edit it with only first name, family name, and the ones you need to change.
}
\examples{
# To add a new student
storeStudents("Jane", "Shepard", "Honours", "Computer Science", "Normandy SR1", "March/2020", "November/2020", FALSE, "Fight the Reapers", "Dr Somebody", "Dr Another Person", "Path/To/Folder", "fileName.csv")

# Otherwise, update them
storeStudents("Jane", "Shepard", completed = TRUE, csvFilePath = "Path/To/Folder", csvFileName = "fileName.csv")

}
