% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-abdc.R
\docType{data}
\name{abdc}
\alias{abdc}
\title{ABDC Journal Quality List}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2685 rows and 7 columns.}
\source{
\url{https://abdc.edu.au/research/abdc-journal-list/}
}
\usage{
data(abdc)
}
\description{
This is a dataset that contains the quality list of rankings of the
Australian Business Deans Council (ABDC). As of 2020-02-19, it contains
the list updated in 2019. You can read more about this list
\href{https://abdc.edu.au/research/abdc-journal-list/}{here}.
}
\details{
Format: a data frame with 2685 observations on the following 7 variables:
\itemize{
\item{\code{journal}: }{Title of the Journal}
\item{\code{publisher}: }{Publishing house}
\item{\code{issn}: }{International Standard Serial Number}
\item{\code{issn_online}: }{ISSN Online - as ISSN, but for the online, rather
than print version}
\item{\code{year_inception}: }{Year the journal started}
\item{\code{field_of_research}: }{Field of Research Code as provided by the
Australian Bureau of Statistics}
\item{\code{rank}: }{In order of best to lowest rank: A*, A, B, or C}
}
}
\examples{
library(dplyr)
abdc \%>\%
  filter(field_of_research == "0104") \%>\%
  arrange(rank)
}
\keyword{datasets}
