\name{Rcademy-package}
\alias{Rcademy-package}
\alias{Rcademy}
\docType{package}
\title{
	www.Rcademy.org: online educational tools for R.
}
\description{
	More details soon
}
\details{
\tabular{ll}{
Package: \tab Rcademy\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-09-05\cr
License: \tab GPL (>= 2) \cr
}
}

\author{
Jonathan Cornelissen
}

\keyword{ package }

