% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcademy-package.R
\docType{package}
\name{rcademy-package}
\alias{rcademy}
\alias{rcademy-package}
\title{rcademy: Tools to assist with academic promotions}
\description{
Ideas and tools to help with preparing documentation for promotions at universities.
}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})

Authors:
\itemize{
  \item Melina Vidoni \email{melina.vidoni@rmit.edu.au} (\href{https://orcid.org/0000-0002-4099-1430}{ORCID})
  \item Belinda Fabian \email{belinda.fabian} (\href{https://orcid.org/0000-0002-5707-4356}{ORCID})
  \item Huong Ly Tong \email{huong-ly.tong@students.mq.edu.au} (\href{https://orcid.org/0000-0002-8462-0105}{ORCID})
  \item Chris Brown \email{Chris.Brown@sydney.edu.au} (\href{https://orcid.org/0000-0003-2506-4871}{ORCID})
  \item Maria Prokofieva \email{Maria.Prokofieva@vu.edu.au} (\href{https://orcid.org/0000-0003-1974-3827}{ORCID})
}

}
\keyword{internal}
