% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_bib}
\alias{read_bib}
\alias{read_pubmed}
\alias{read_scholar}
\alias{read_orcid}
\title{Read bibliographies}
\usage{
read_bib(filename)

read_pubmed(query)

read_scholar(id)

read_orcid(id)
}
\arguments{
\item{filename}{The filename of a bib file (i.e., in BibTeX format)}

\item{query}{A character string containing a search query to pass to PubMed}

\item{id}{A character string specifying the Google Scholar ID or Orcid ID}
}
\value{
A tibble containing one row per publication. Columns include title, authors, year, journal, etc.
}
\description{
Create tables of publications from bib files, or from PubMed, Orcid or Google Scholar
}
\examples{

\dontrun{

mypubs <- read_bib("mypubs.bib")
mypubs <- read_pubmed("Huang Ly Tong")
mypubs <- read_scholar("EUdX6oIAAAAJ")
mypubs <- read_orcid("0000-0002-8462-0105")
}

}
\author{
Rob J Hyndman
}
